pyproject.toml
[tool.poetry]
name = "constitutionai"
version = "0.1.0"
description = ""
authors = ["Saikat Kumar Dey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
embedchain = {extras = ["opensource"], version = "^0.0.59"}
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"
pandas = "2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
